Established Preprocessing PipleLine Connection
The job 'data_intake' reads POSTGRES into a pandas df and then to a spark df. Here, a spark object manipulates the dataframe to transform and clean for analysis/reference in a later stage. This then saved to a csv with Pandas and is read into the preprocessor by spark from the dedicated dir...clean and ready to go.